pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "cap"
version = "0.0.1"
description = "End to end ML project about predicting credit approval."
authors = [
  { name = "Gustavo Hern\u00e1ndez Angeles" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "awscli",
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "seaborn>=0.13.2",
    "tqdm",
    "typer",
]
requires-python = "~=3.12.0"


[tool.ruff]
line-length = 99
src = ["cap"]
include = ["pyproject.toml", "cap/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["cap"]
force-sort-within-sections = true

[dependency-groups]
dev = [
    "prefect>=3.4.7",
    "pytest>=8.4.1",
]

